Return batch data as a hash of URLs/data instead of just data.
When a parse fails, Diffbot may silently discard the particular response
for the URL in the batch request. So it's necessary to keep a hash of
the URLs and match them up to the response by the 'relative url' sent in
the request.

For URLs that fail, the hash entry returned by the batch method will be nil.